requirements.txt
firebase-admin==6.1.0
google-cloud-pubsub==2.14.1
google-cloud-speech==2.17.3
google-cloud-storage==2.7.0
ffmpeg-python==0.2.0